user: tmux.service: cleanup environment prior to creating a session
`ExecStartPost=` is too late to unset systemd-specific environment
variables from the server global environment as by that point
the session is already started (which spawns a shell automatically).

Instead, clean up the server global environment prior to issuing a
`new-session` command as part of `ExecStart=`. To avoid cluttering
the command line too much, create and source a configuration file
(shipped with the unit) that contains the `set-environment` commands.