scripts: import-profile.sh: override bogus $SHELL for any UID
For reasons that elude me once again, the user session on the Fedora
Live DVD starts with a bogus SHELL=/bin/bash in its environment.

Extend the previous hack to any UID, and wrap it in an extra check
to only perform the override if we did in fact end up with a bogus shell
after reading the system and user profile files.